LayoutTests/http/tests/xmlhttprequest/access-control-preflight-sync-method-denied-expected.txt

CONSOLE MESSAGE: line 34: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php?test=access-control-preflight-sync-method-denied&state=method. Method DELETE is not allowed by Access-Control-Allow-Methods.
PASS: Request successfully blocked.
